pyproject.toml
[project]
name = "tgdlbot"
version = "0.1.0"
description = "Telegram bot for downloading YouTube videos"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=21.0",
    "yt-dlp>=2024.12.6",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
